test(googlechat): split thread reply coverage
